+++ Escuela/Sexto/listita/app/src/main/java/com/example/listita/calculos.java
@@ -2,27 +2,27 @@ package com.example.listita;
 
 public class calculos
 {
-    public double suma (int num1, int num2)
+    public double suma (double num1, double num2)
     {
-        int total = 0;
+        double total = 0;
         total = num1 + num2;
         return total;
     }
-    public double resta (int num1, int num2)
+    public double resta (double num1, double num2)
     {
-        int total = 0;
+        double total = 0;
         total = num1 - num2;
         return total;
     }
-    public double multi (int num1, int num2)
+    public double multi (double num1, double num2)
     {
-        int total = 0;
+        double total = 0;
         total = num1 * num2;
         return total;
     }
-    public double divide (int num1, int num2)
+    public double divide (double num1, double num2)
     {
-        int total = 0;
+        double total = 0;
         total = num1 / num2;
         return total;
     }
